# Properly cache team checksums